+++ budget/update/service/models/request_model.go
@@ -2,8 +2,8 @@ package models
 
 // Create struct to hold info about new item
 type RequestModel struct {
-	Pk       *string `validate:"required" json:"pk"`
-	Sk       *string `validate:"required" json:"sk"`
-	Planned  *int64  `validate:"required" json:"planned"`
-	Category *string `validate:"required" json:"category"`
+	Pk       *string  `validate:"required" json:"pk"`
+	Sk       *string  `validate:"required" json:"sk"`
+	Planned  *float64 `validate:"required" json:"planned"`
+	Category *string  `validate:"required" json:"category"`
 }
