+++ dynamodb-stream/debt-link-trigger/update/service/models/debt_rule.go
@@ -2,10 +2,10 @@ package models
 
 // Create struct to hold info about retrieved item
 type DebtRule struct {
-	Pk                *string `json:"pk"`
-	Sk                *string `json:"sk"`
-	CreationDate      *string `json:"creation_date"`
-	TransactionName   *string `json:"transaction_name"`
-	TransactionAmount *int64  `json:"transaction_amount"`
-	DebtId            *string `json:"debt_id"`
+	Pk                *string  `json:"pk"`
+	Sk                *string  `json:"sk"`
+	CreationDate      *string  `json:"creation_date"`
+	TransactionName   *string  `json:"transaction_name"`
+	TransactionAmount *float64 `json:"transaction_amount"`
+	DebtId            *string  `json:"debt_id"`
 }
