+++ transactions/update/service/models/query_parameter.go
@@ -2,7 +2,7 @@ package models
 
 // Create struct to hold info about new item
 type QueryParameter struct {
-	Amount      *int64    `validate:"required" json:":a"`
+	Amount      *float64  `validate:"required" json:":a"`
 	Description *string   `validate:"required" json:":d"`
 	CategoryId  *string   `validate:"required" json:":c"`
 	Tags        *[]string `validate:"required" json:":t"`
