+++ transactions/update/service/models/request_model.go
@@ -4,7 +4,7 @@ package models
 type RequestModel struct {
 	Pk          *string   `validate:"required" json:"pk"`
 	Sk          *string   `validate:"required" json:"sk"`
-	Amount      *int64    `validate:"required" json:"amount"`
+	Amount      *float64  `validate:"required" json:"amount"`
 	Description *string   `validate:"required" json:"description"`
 	CategoryId  *string   `validate:"required" json:"category_id"`
 	Tags        *[]string `validate:"required" json:"tags"`
